+++ modules/fwstate/api/fwstate_cp.h
@@ -87,6 +87,6 @@ fwstate_config_cursor_create(
 	fwstate_cursor_t *cursor,
 	bool is_ipv6,
 	uint32_t layer_index,
-	uint32_t index,
+	int64_t index,
 	bool include_expired
 );
